Tests and refactoring of RecommenderAPIProvider (#112)
The API was refactored so that it is not singleton (this allowed better
testing) + now it is created with URL, token and user key as constructor
arguments. This allowed to leave out these parameters from method calls.
Also, the whole provider instance is passed to the analyzer job so that
it can call the methods directly with no need to retrieve user key or
url from the storage. :

The 3scale method was extracted into separate class and tests were
created.